+++ src/tests/scenarios/diff.rs
@@ -3,4 +3,5 @@ use super::*;
 mod changed_ranges;
 mod cli;
 mod filter;
+mod hook;
 mod parser;